41 implement stateful widget instead of widget

+++ Justfile
@@ -20,9 +20,9 @@ windows-test:
 
 
 # Generate HTML test coverage report in target/coverage/html
-# Requires cargo-binutils and grcov to be installed
+# Requires cargo-binutils, grcov and rustup llvm-tools to be installed
 cov:
-    rm -r target/coverage/html
+    [ -d target/coverage/html ] && rm -r target/coverage/html || true
     CARGO_INCREMENTAL=0 RUSTFLAGS='-Cinstrument-coverage' LLVM_PROFILE_FILE='cargo-test-%p-%m.profraw' cargo test
     grcov . --binary-path ./target/debug/deps/ -s . -t html --branch --ignore-not-existing --ignore '../*' --ignore "/*" -o target/coverage/html
     rm *profraw
